Fix NessUUID bug with decode
This is a correctness bug that never manifests in a well-formed UUID
because no segment is even close to 16 characters. But it is entirely
legal to have a nibble being shifted into the topmost nibble that
makes the resulting long negative. So the check in decode() was wrong
and it is much easier to simply check on maxlength (no string can have
more than 16 chars (which is 64 bits).

Also fix the various bad tests for toString().